[IMP] web: convert journal dashboard graph field tests to hoot
task-3705027
